backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
httpx==0.27.2
openai==1.51.0
supabase==2.9.0
pgvector==0.3.3
psycopg2-binary==2.9.9
youtube-transcript-api==0.6.2
PyPDF2==3.0.1
pdfplumber==0.11.4
langchain==0.3.1
langchain-openai==0.2.1
langchain-community==0.3.1
python-dotenv==1.0.1
pydantic==2.9.2
tiktoken==0.7.0
yt-dlp==2024.9.27
sentence-transformers==3.2.1
